zapcore: breakout sampler_test from facility_test
